Install a ConditionFirstBoot service for /root/.bashrc
This is cleaner than doing it in the `cloud.ks`; it works for bare
metal too.  If you do a full factory reset it comes back, etc.

The reason I'm doing this now specifically is that I want to be
able to make a change to `/etc/skel` and have that consistently
reflected for root too.