Split off ootpa fork
We really want separate release packages for our separate
streams to help distinguish them; there's a pretty huge difference
at the OS level.

Note that this also fixes a rather large bug in that we were missing
the `99-default-disable.preset`.  I'll do another PR to add that
to the maipo `master` too.