Unify build scripts
We have separate branches now, so there's no need to have the split base
and ootpa approach. Let's just have a single `build.sh` script to keep
things simple.

This also fixes the issue of `NAME=` and `VERSION=` appearing twice in
`/etc/os-release` (with only the latter having the right values). This
is fine when sourcing the file in a script, but it can be confusing for
humans and parsers.